factored out into utils
